added code for login

+++ Learnstream/backend/src/utils/cloudinary.js
@@ -1,7 +1,9 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
-
-
+import dotenv from "dotenv"
+dotenv.config({
+    path: './.env'
+})
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
   api_key: process.env.CLOUDINARY_API_KEY, 
@@ -21,7 +23,10 @@ const uploadOnCloudinary = async (localFilePath) => {
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        else console.warn(`File not found at file path${localFilePath}`);
+        
         return null;
     }
 }
